requirements.txt
numpy
scipy
svg.path >= 4
dxfgrabber
shapely
